Update code to reflect showMessageBox not blocking
